change db model to reflect current data metric schema change
